refactor: moved elements & layouts folder to components folder
