+++ resources/js/Pages/components/elements/card/DefaultCard.vue
