+++ resources/js/Pages/components/elements/tables/index.vue
