+++ resources/js/Pages/components/elements/tabview/TabViewPage.vue
