+++ resources/js/Pages/components/layouts/SubscriptionsLayout.vue
@@ -1,10 +1,10 @@
 <script setup>
 import { onMounted } from 'vue';
-import DashboardLayout from '@/layouts/DashboardLayout.vue';
+import DashboardLayout from '@/components/layouts/DashboardLayout.vue';
 
 import { useNavMainPlatform } from '@/stores/navigation_menu_item';
 import { storeToRefs } from 'pinia';
-import TabViewPage from '@/elements/tabview/TabViewPage.vue';
+import TabViewPage from '@/components/elements/tabview/TabViewPage.vue';
 import { useSubscriptionTabs } from '@/stores/subscriptions_tabs';
 
 const getNavMainPlatform = useNavMainPlatform()
