+++ resources/js/Pages/views/dashboard/central_page/Overview.vue
@@ -5,8 +5,8 @@ import { useNavMainPlatform } from "@/stores/navigation_menu_item";
 import { useMenuUser } from "@/stores/menu_dropdown_user";
 import { storeToRefs } from 'pinia';
 import Divider from 'primevue/divider';
-import Card from '@/elements/card/DefaultCard.vue';
-import DashboardLayout from '@/layouts/DashboardLayout.vue';
+import Card from '@/components/elements/card/DefaultCard.vue';
+import DashboardLayout from '@/components/layouts/DashboardLayout.vue';
 import TenantLatestSection from '@dashboardCentralPageComponents/mainMenu/tenantLatest_section.vue';
 import LatestTransactionSubscriptionSection from '@dashboardCentralPageComponents/mainMenu/LatestTransactionSubscription_section.vue';
